backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]
aiosqlite
httpx
opik
hume
python-multipart
# RAG dependencies
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.0
pyyaml>=6.0
python-frontmatter>=1.0.0